+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-1104] Changed `/api/data_structures/{id}/versions/{version}` to also return versions
+
 ## [2.9.3] 2018-12-04
 
 ### Added
 
-- [TD-1104] API endpoint `/api/data_structures/{id}/version/{version}` to read specific version of a data structure
+- [TD-1104] API endpoint `/api/data_structures/{id}/versions/{version}` to read specific version of a data structure
 
 ## [2.9.2] 2018-12-04